+++ python/src/iceberg/expressions/base.py
@@ -17,9 +17,10 @@
 from abc import ABC, abstractmethod
 from enum import Enum, auto
 from functools import reduce
-from typing import Any, Generic, TypeVar, Optional
+from typing import Any, Generic, TypeVar
 
 from iceberg.files import StructProtocol
+from iceberg.schema import Accessor
 from iceberg.types import NestedField, Singleton
 
 T = TypeVar("T")
@@ -264,45 +265,6 @@ class AlwaysFalse(BooleanExpression, Singleton):
 
     def __str__(self) -> str:
         return "false"
-
-
-# TODO: Fokko check if a dataclass is nice here
-class Accessor:
-    """An accessor for a specific position in a container that implements the StructProtocol"""
-
-    def __init__(self, position: int, inner: Optional["Accessor"] = None):
-        self._position = position
-        self._inner = inner
-
-    def __str__(self):
-        return f"Accessor(position={self._position},inner={self._inner})"
-
-    def __repr__(self):
-        return self.__str__()
-
-    def __eq__(self, other):
-        return other and self.position == other.position
-
-    @property
-    def position(self):
-        """The position in the container to access"""
-        return self._position
-
-    @property
-    def inner(self) -> Optional["Accessor"]:
-        """The inner object in the case of a struct"""
-        return self._inner
-
-    def get(self, container: StructProtocol) -> Any:
-        """Returns the value at self.position in `container`
-
-        Args:
-            container(StructProtocol): A container to access at position `self.position`
-
-        Returns:
-            Any: The value at position `self.position` in the container
-        """
-        return container.get(self.position)
 
 
 class BoundReference:
